+++ data/i18n/hydrogen_es.ts
@@ -3117,6 +3117,18 @@ Debería funcionar correctamente mientras utilices el GMRockKit, y no uses tresi
     </message>
 </context>
 <context>
+    <name>MidiControlButton</name>
+    <message>
+        <source>MIDI mapping and monitoring</source>
+        <extracomment>Tooltip of the MIDI button in the main toolbar.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>No MIDI driver present!</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>MidiControlDialog</name>
     <message>
         <source>Incoming</source>