+++ data/i18n/hydrogen_fr.ts
@@ -3118,6 +3118,18 @@ L&apos;exportation LilyPond est une fonctionnalité expérimentale.
     </message>
 </context>
 <context>
+    <name>MidiControlButton</name>
+    <message>
+        <source>MIDI mapping and monitoring</source>
+        <extracomment>Tooltip of the MIDI button in the main toolbar.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>No MIDI driver present!</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>MidiControlDialog</name>
     <message>
         <source>Incoming</source>